+++ src/main/java/net/backslot/mixin/PlayerInventoryMixin.java
@@ -5,10 +5,7 @@ import java.util.List;
 
 import com.google.common.collect.ImmutableList;
 
-import org.spongepowered.asm.mixin.Final;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Mutable;
-import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.*;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -29,23 +26,13 @@ public abstract class PlayerInventoryMixin implements Inventory {
     @Mutable
     private List<DefaultedList<ItemStack>> combinedInventory;
     @Shadow
-    @Final
-    @Mutable
-    public DefaultedList<ItemStack> main;
-    @Shadow
-    @Final
-    @Mutable
-    public DefaultedList<ItemStack> armor;
-    @Shadow
-    @Final
-    @Mutable
-    public DefaultedList<ItemStack> offHand;
-    @Shadow
     @Mutable
     @Final
     public PlayerEntity player;
 
+    @Unique
     private DefaultedList<ItemStack> backSlot;
+    @Unique
     private DefaultedList<ItemStack> beltSlot;
 
     public PlayerInventoryMixin(PlayerEntity player) {